+++ src/main/java/communication/messages/Message.java
@@ -1,5 +1,4 @@
 package communication.messages;
 
 public class Message {
-
 }