+++ src/main/java/controller/server/GameServer.java
@@ -1,18 +1,5 @@
 package controller.server;
 
-import com.squareup.moshi.JsonAdapter;
-import com.squareup.moshi.Moshi;
-import communication.messages.JoinGameRequest;
-
-import java.io.IOException;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.util.ArrayList;
-import java.util.List;
-
-import communication.messages.Message;
-import org.json.JSONObject;
-
 public class GameServer {
 
 }